add double numbers to the calculator

+++ src/test/java/calculator/TestStringToExpression.java
@@ -16,55 +16,55 @@ public class TestStringToExpression {
     public void StringToExpressionPositiveNumbersTest() throws Exception{
         
         Expression e = StringToExpression.parseStringTExpression("{(4 * 3) + [2 * 5]} / 5)))");
-        int result = c.eval(e); 
+        int result = c.eval(e).intValue(); 
         assertEquals(4, result);
 
 
         e = StringToExpression.parseStringTExpression("50 - 3 * 4 + 10 - 2 * 5 + 8 * 2");
-        result = c.eval(e); 
+        result = c.eval(e).intValue(); 
         assertEquals(54, result);
 
         e = StringToExpression.parseStringTExpression("(10 + 5) * 3 - (8 - 2) * 4 + 6");
-        result = c.eval(e); 
+        result = c.eval(e).intValue(); 
         assertEquals(27, result);
 
         e = StringToExpression.parseStringTExpression("((10 + 2) * 3 - 5) + ((8 - 4) * 6) - (7 * 2) + 9");
-        result = c.eval(e); 
+        result = c.eval(e).intValue(); 
         assertEquals(50, result);
 
 
         e = StringToExpression.parseStringTExpression("10 * 3 - 5 * 2");
-        result = c.eval(e); 
+        result = c.eval(e).intValue(); 
         assertEquals(20, result);
 
 
         e = StringToExpression.parseStringTExpression("5 + 10 * 2 - 3");
-        result = c.eval(e); 
+        result = c.eval(e).intValue(); 
         assertEquals(22, result);
 
 
         e = StringToExpression.parseStringTExpression("(5 + 3) * 2 - (4 - 2) * 3");
-        result = c.eval(e); 
+        result = c.eval(e).intValue(); 
         assertEquals(10, result);
 
         e = StringToExpression.parseStringTExpression("1000 - 5 * 200 + 50");
-        result = c.eval(e); 
+        result = c.eval(e).intValue(); 
         assertEquals(50, result);
 
         e = StringToExpression.parseStringTExpression("((10 + 5) * (3 - 2) + (8 * 2)) - (6 * (4 - 2)) + 9");
-        result = c.eval(e); 
+        result = c.eval(e).intValue(); 
         assertEquals(28, result);
 
         e = StringToExpression.parseStringTExpression("(3 + 5) * (9 - 2) ");
-        result = c.eval(e); 
+        result = c.eval(e).intValue(); 
         assertEquals(56, result);
 
         e = StringToExpression.parseStringTExpression("1+2");
-        result = c.eval(e); 
+        result = c.eval(e).intValue(); 
         assertEquals(3, result);
 
         e = StringToExpression.parseStringTExpression("3*7");
-        result = c.eval(e); 
+        result = c.eval(e).intValue(); 
         assertEquals(21, result);
      
     }
@@ -194,25 +194,25 @@ public class TestStringToExpression {
     public void StringToExpressionNegativeNumbersTest() throws Exception{
         
         Expression e = StringToExpression.parseStringTExpression("4 -- 2");
-        int result = c.eval(e); 
+        int result = c.eval(e).intValue(); 
         assertEquals(6, result);
 
 
         e = StringToExpression.parseStringTExpression("4 - - - 2");
-        result = c.eval(e); 
+        result = c.eval(e).intValue(); 
         assertEquals(2, result);
 
         e = StringToExpression.parseStringTExpression("-2 + 4");
-        result = c.eval(e); 
+        result = c.eval(e).intValue(); 
         assertEquals(2, result);
 
         e = StringToExpression.parseStringTExpression("-2 + - 2");
-        result = c.eval(e); 
+        result = c.eval(e).intValue(); 
         assertEquals(-4, result);
 
 
         e = StringToExpression.parseStringTExpression("(1-(-(-(-1))))");
-        result = c.eval(e); 
+        result = c.eval(e).intValue(); 
         assertEquals(2, result);
      
     }
